Clarify app-local test boundaries
