+++ apps/favn_authoring/test/assets/graph_planner_parity_test.exs
@@ -1,4 +1,4 @@
-defmodule Favn.Assets.GraphPlannerParityTest do
+defmodule FavnAuthoring.Assets.GraphPlannerParityTest do
   use ExUnit.Case, async: false
 
   alias Favn.Assets.GraphIndex
