+++ apps/favn_runner/test/favn_runner_test.exs
@@ -6,11 +6,8 @@ defmodule FavnRunnerTest do
   alias Favn.Manifest.Asset
   alias Favn.Manifest.Graph
   alias Favn.Manifest.Version
-  alias FavnTestSupport.Fixtures
 
   setup do
-    Fixtures.compile_fixture!(:runner_assets)
-
     manifest_version = "mv_" <> Base.encode16(:crypto.strong_rand_bytes(8), case: :lower)
 
     manifest =
@@ -39,17 +36,17 @@ defmodule FavnRunnerTest do
     %{version: version}
   end
 
-  test "runs a shared fixture asset through runner execution boundary", %{version: version} do
-    fixture_ref = {Favn.Test.Fixtures.Assets.Runner.RunnerAssets, :base}
+  test "runs a local plain Elixir asset through runner execution boundary", %{version: version} do
+    fixture_ref = {FavnRunnerTest.PlainElixirAsset, :asset}
 
     fixture_manifest =
       build_manifest([
         %Asset{
           ref: fixture_ref,
           module: elem(fixture_ref, 0),
-          name: :base,
+          name: :asset,
           type: :elixir,
-          execution: %{entrypoint: :base, arity: 1}
+          execution: %{entrypoint: :asset, arity: 1}
         }
       ])
 
@@ -139,6 +136,13 @@ defmodule FavnRunnerTest do
       metadata: %{}
     }
   end
+end
+
+defmodule FavnRunnerTest.PlainElixirAsset do
+  alias Favn.Run.Context
+
+  @spec asset(Context.t()) :: {:ok, map()}
+  def asset(%Context{} = ctx), do: {:ok, %{partition: ctx.params[:partition]}}
 end
 
 defmodule FavnRunnerTest.ElixirAsset do
