+++ docs/structure/favn_core.md
@@ -8,6 +8,8 @@ Code:
 
 Tests:
 - `apps/favn_core/test/`
+- App-local tests must use only `favn_core` and declared test support; authoring
+  DSL parity tests live in `favn_authoring`.
 
 Use when changing manifest generation/serialization, graph planning, dependency
 inference, windows, schedules, runtime config refs, backfill range resolution, or
